Support named connections
